requirements.txt
tk
customtkinter==5.2.0
Pillow==8.4.0
Pillow==7.0.0
Pillow==10.0.1
